made changes to code base for easier multi-sig usage
