+++ .env.example
@@ -4,5 +4,8 @@ DEPLOY_ON = localnet
 # private key of the deployer
 DEPLOYER_KEY = 0x9e36456bda63e8fb4f33cd40ca6fa0afc14dbdd58de1071eb52b77a7231fb971
 
+# seed phrase for deployer (optional) if DEPLOYER_KEY is provided, DEPLOYER_PHRASE won't be used
+DEPLOYER_PHRASE = parent dust canal tourist grape space repeat explain rather cushion permit stadium
+
 # type of wallet. only supported Secp256k1 and ED25519. defaults to Secp256k1
 WALLET_SCHEME = ED25519
