+++ .vscode/settings.json
@@ -1,9 +1,11 @@
 {
     "cSpell.words": [
         "bignumber",
+        "interactor",
         "Interactor",
         "Keypair",
         "localnet",
+        "mainnet",
         "mysten",
         "Secp"
     ]
