+++ deployment.json
@@ -1,16 +1,9 @@
 {
-  "devnet": {
-    "TreasuryCap": "0x55187852f1db4e7436097bafcf0bb8fcf08a8388336d238bb7f71999799bdab7",
-    "Package": "0xb4dd272a088d26ca54f8e57950a3ea4fd48984a98065fe3267d3ea3d720803f5",
-    "CoinMetadata": "0xbfd296c9f6020a9f0c8dff42586eae77d329222980e3dc191d2467f8324b534f",
-    "UpgradeCap": "0xf15aaac3364b959e3812b4f736827ea6cdeb04430ec036772d1f591905d21833",
-    "BasePackage": "0xb4dd272a088d26ca54f8e57950a3ea4fd48984a98065fe3267d3ea3d720803f5"
-  },
   "testnet": {
-    "UpgradeCap": "0x19a5f3e077c0edca4c59cc3d331bfa46f702384f72cfba14d22cd942013ac927",
-    "CoinMetadata": "0x9cf6f9d95bb557a3970064b935c91ca51871e6a53a11262058c627f1dccdd81b",
-    "TreasuryCap": "0xa5f0848178cb408fb53a2caa8e29116d370b33ef1236274cdeb63122fed29604",
-    "Package": "0xe70937cccc82ea90d8313d8e198b5ce9162bc33a78fc14f5b0ba8c912848883d",
-    "BasePackage": "0xe70937cccc82ea90d8313d8e198b5ce9162bc33a78fc14f5b0ba8c912848883d"
+    "UpgradeCap": "0xa69899567cf10f4a30cb5b76a639ee07a3ad4bdba4bfbc464ea1526f17854da5",
+    "CoinMetadata": "0x57f09935ad594290a048406bf41a9c6e53d1146e01cee458155ad2c16d080fb5",
+    "TreasuryCap": "0xf735f09787b9f4e5da7f7725fd980e3b3d07ecffb5521bf08bb5d922c3e423c5",
+    "Package": "0x0a04404827683b5ca5974f55a4f4a650755ef3289f202fb8f20551ae2d86dec6",
+    "BasePackage": "0x0a04404827683b5ca5974f55a4f4a650755ef3289f202fb8f20551ae2d86dec6"
   }
 }
