+++ multi-sig-wallet.json
@@ -1,0 +1,21 @@
+{
+  "multisigAddress": "0xca9e2056044e12ddccc762e8f0661a8a60ee19e67f95a2f000fa532a63c3afa4",
+  "multisig": [
+    {
+      "address": "0x80c3d285c2fe5ccacd1a2fbc1fc757cbeab5134f1ef1e97803fe653e041c88aa",
+      "publicBase64Key": "AQLMsPtNd9cWgIl1Ii3KvOfHxc90n9Fpe9ledKo3M2dtmQ==",
+      "weight": 1
+    },
+    {
+      "address": "0x2183df5aaf6366e5445c95fa238fc223dbbda54b7c363680578b435f657f1a29",
+      "publicBase64Key": "AQLUiOOWb0TKC0huixPCmWIWVyQ8C31YzcV44UE/qkhUpw==",
+      "weight": 1
+    },
+    {
+      "address": "0xed2bb2ae1330a3abee7794e659add176b827e13532b31074ad01330df2d5c843",
+      "publicBase64Key": "AQMVXh1oKkbck+bUGysJwqtzxgpHw9S32DAdLMNY9ifs3g==",
+      "weight": 1
+    }
+  ],
+  "threshold": 2
+}
