+++ multisig/execTx.ts
@@ -1,16 +1,16 @@
-import { SUI_CLIENT } from "../scripts/utils";
-import { MULTI_SIG_WALLET, combineMSTx } from "./utils";
+import { MULTI_SIG_WALLET, SUI_CLIENT } from "../env";
+import { combineMSTx } from "../utils";
 
 async function main() {
   const txBytes =
-    "AAABACDKniBWBE4S3czHYujwZhqKYO4Z5n+VovAA+lMqY8OvpAIEAbwHoRzrCwYAAAAKAQAMAgwkAzBDBHMOBYEBkAEHkQLyAQiDBGAG4wTAAQqjBgUMqAZjAAcBEAIJAhYCFwIYAAACAAEDBwEAAAIBDAEAAQICDAEAAQIEDAEAAQQFAgAFBgcAAAwAAQAADQIBAAAIAwEAARQHCAEAAggUEgEAAgoKCwECAg4TAQEAAhUREgEAAxEHAQEMAxIQAQEMBBMNDgAFDwUGAAMGBQkIDAkPBwkGCQQJAggABwgFAAQHCwQBCAADBQcIBQIHCwQBCAALAgEIAAILAwEIAAsEAQgAAQoCAQgGAQkAAQsBAQkAAQgABwkAAgoCCgIKAgsBAQgGBwgFAgsEAQkACwMBCQABCwMBCAABBggFAQUBCwQBCAACCQAFAQYLBAEJAAEDBAcLBAEJAAMFBwgFAgcLBAEJAAsCAQkABEJMVUUEQ29pbgxDb2luTWV0YWRhdGEGT3B0aW9uC1RyZWFzdXJ5Q2FwCVR4Q29udGV4dANVcmwEYmx1ZQRidXJuBGNvaW4PY3JlYXRlX2N1cnJlbmN5C2R1bW15X2ZpZWxkBGluaXQEbWludBFtaW50X2FuZF90cmFuc2ZlchVuZXdfdW5zYWZlX2Zyb21fYnl0ZXMGb3B0aW9uFHB1YmxpY19mcmVlemVfb2JqZWN0D3B1YmxpY190cmFuc2ZlcgZzZW5kZXIEc29tZQx0b3RhbF9zdXBwbHkIdHJhbnNmZXIKdHhfY29udGV4dAN1cmwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAIDCAAAZKeztuANAwgBAAAAAAAAAAoCBQRCTFVFCgIIB0JsdWVmaW4KAm5tQmx1ZWZpbiBmb3VuZGF0aW9uIGNvaW4gZWFybmVkIGJ5IHRyYWRpbmcgb24gdGhlIEJsdWVmaW4gcHJvdG9jb2wgYW5kIHVzZWQgdG8gcGFydGFrZSBpbiBnb3Zlcm5hbmNlIHByb3Bvc2FscwoCJSRodHRwczovL2JsdWVmaW4uaW8vaW1hZ2VzL3NxdWFyZS5wbmcAAgELAQAAAAAEFAsAMQkHAgcDBwQHBRELOAAKATgBDAIMAwsCOAILAwsBLhEKOAMCAQEEAAEVCgAuOAQKARYHACUECQUPCwABCwMBBwEnCwALAQsCCwM4BQICAQQAAQULAAsBOAYBAgACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAgEBAwAAAAABAADKniBWBE4S3czHYujwZhqKYO4Z5n+VovAA+lMqY8OvpAFRaedO3f2NMJtpb+9x11fZ8VSeY7gNyd/bMNhUbgMQnC1DCgAAAAAAILKQvudhJSYPsLleMm+dOCSjSl+2xODDCZ4vOCJGXCWkyp4gVgROEt3Mx2Lo8GYaimDuGeZ/laLwAPpTKmPDr6ToAwAAAAAAAEBTGwEAAAAAAA==";
+    "AAACAD8lTi/HGBbFYtlG3fXIOtCBSD0ToaOPwyhPpqk/XRVBFzhHxIAAAAAIAiYVrYhSk9ix2B/xZsJEx9uufR+SpFe9oRjL+NFRe4DACAhg99ar2Nm5URclfojj8Ij272lS3w2NoBXi0NfZX8aKQEBAQEAAAEBAAvy0ItWR+mLUi0uwh9dPDpvG16YEUSWmcUMzYjWG4Z7AXgD1Skg3aIPNhyhCnk3TQfPvovUxt5e69Sh3JXfWcW8XeEfEgAAAAAgDnnhqnS0XLq0vUsNeYSVjWUnAMzCdfwS5RdyazcfL1IL8tCLVkfpi1ItLsIfXTw6bxtemBFElpnFDM2I1huGe+4CAAAAAAAAlDYmAAAAAAAA";
 
   const signatures = [
-    "AU9rtnxetvEsytwb6AlDruNfm06/O58Qmt2k055fD9GkVw/boyizmaTccyDrfO1kkKHOo0/B+fmhSGHofYQFsPgCzLD7TXfXFoCJdSItyrznx8XPdJ/RaXvZXnSqNzNnbZk=", // wallet 1
+    "AamhpafUi6JIe37a/sbaphk7A04WdUM2Hsvh5MnnOWKCQczmZ6fQSMT20fe1oCiK5PhvOVIxLul5+swvFGrrNMQDSQv7fZB1KB4AqYYUq/Fix2vIm+UcJdbKzTAFwkIP8gk=", // wallet 1
     "ActcHr+nve4CAGLIL7/Kd0aRQKvkdYnRCLbSfzyDgnWVceqKcD88ujISRR54JX0+3knT0+OvP139PaJo4oiou3wC1Ijjlm9EygtIbosTwpliFlckPAt9WM3FeOFBP6pIVKc=", // wallet 2
     "AZG4ONiIUAnbchwwU7Y+m7HrcZtPogL8+TJnftirRSvxFfCM0wS7XXV47/Zdp7V+ofksSS42pFwG5fqhFqCNdsUDFV4daCpG3JPm1BsrCcKrc8YKR8PUt9gwHSzDWPYn7N4=", // wallet 3
   ];
-
+  
   const msSign = combineMSTx(signatures, MULTI_SIG_WALLET);
   console.log(msSign);
