+++ scripts/build.ts
@@ -2,6 +2,7 @@
 
 import { execSync } from "child_process";
 import path from "path";
+
 async function main() {
   const pkgPath = path.join(path.resolve(__dirname), "../");
   execSync(`sui move build --dump-bytecode-as-base64 --path ${pkgPath}`, {
